requirements.txt
connexion[swagger-ui]<3
flask>=2.3.3
Werkzeug>=2.0
gunicorn>=21.2.0
psycopg2-binary==2.9.9
sqlalchemy